Chore: cleaning Detailed Questions page
Split codes in different components and imported them
