+++ app/Components/forDetailed/fetchQuestion.ts
@@ -1,0 +1,15 @@
+export async function fetchQuestion(id: string) {
+    const res = await fetch(`https://nunu29.pythonanywhere.com/questions/${id}/`);
+  
+    if (res.status === 404) {
+      return null;
+    }
+  
+    if (!res.ok) {
+      throw new Error("Failed to fetch data");
+    }
+  
+    const data = await res.json();
+    return data;
+  }
+
